requirements.txt
fastapi[standard]
click>=8.0
openai
pylint
pytest
bandit
pytest-json-report
python-dotenv